refactor: move handlers to pkg/handlers/lib directory structure
Reorganized handler packages to be under a lib subdirectory:
- pkg/handlers/homebrew → pkg/handlers/lib/homebrew
- pkg/handlers/path → pkg/handlers/lib/path
- pkg/handlers/shell → pkg/handlers/lib/shell
- pkg/handlers/install → pkg/handlers/lib/install
- pkg/handlers/symlink → pkg/handlers/lib/symlink

Updated all import paths across the codebase to reflect the new structure.
All tests pass successfully with the new organization.
