+++ pkg/core/init.go
@@ -5,11 +5,11 @@ import (
 	"github.com/arthur-debert/dodot/pkg/logging"
 
 	// Import all handler packages to register their factories
-	_ "github.com/arthur-debert/dodot/pkg/handlers/homebrew"
-	_ "github.com/arthur-debert/dodot/pkg/handlers/install"
-	_ "github.com/arthur-debert/dodot/pkg/handlers/path"
-	_ "github.com/arthur-debert/dodot/pkg/handlers/shell"
-	_ "github.com/arthur-debert/dodot/pkg/handlers/symlink"
+	_ "github.com/arthur-debert/dodot/pkg/handlers/lib/homebrew"
+	_ "github.com/arthur-debert/dodot/pkg/handlers/lib/install"
+	_ "github.com/arthur-debert/dodot/pkg/handlers/lib/path"
+	_ "github.com/arthur-debert/dodot/pkg/handlers/lib/shell"
+	_ "github.com/arthur-debert/dodot/pkg/handlers/lib/symlink"
 )
 
 // Initialize sets up the core system by:
