+++ pkg/handlers/lib/homebrew/homebrew_clear_test.go
