+++ pkg/handlers/lib/install/install-template.txt
