+++ pkg/handlers/lib/install/integration_test.go
@@ -1,4 +1,4 @@
-package homebrew_test
+package install_test
 
 import (
 	"fmt"
@@ -6,7 +6,7 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/arthur-debert/dodot/pkg/handlers/homebrew"
+	"github.com/arthur-debert/dodot/pkg/handlers/lib/install"
 	"github.com/arthur-debert/dodot/pkg/operations"
 	"github.com/arthur-debert/dodot/pkg/types"
 	"github.com/stretchr/testify/assert"
@@ -58,28 +58,25 @@ func (m *MockSimpleDataStore) ListHandlerSentinels(pack, handlerName string) ([]
 	return args.Get(0).([]string), args.Error(1)
 }
 
-func TestHomebrewHandler_OperationIntegration(t *testing.T) {
-	// This test verifies the homebrew handler works with the operation system
+func TestInstallHandler_OperationIntegration(t *testing.T) {
+	// This test verifies the install handler works with the operation system
 
-	// Create test Brewfile
+	// Create test script
 	tempDir := t.TempDir()
-	brewfileContent := `# Test Brewfile
-brew "git"
-cask "visual-studio-code"
-`
-	brewfilePath := filepath.Join(tempDir, "Brewfile")
-	err := os.WriteFile(brewfilePath, []byte(brewfileContent), 0644)
+	scriptContent := "#!/bin/bash\necho 'Installing test pack'\n"
+	scriptPath := filepath.Join(tempDir, "install.sh")
+	err := os.WriteFile(scriptPath, []byte(scriptContent), 0755)
 	require.NoError(t, err)
 
 	// Create simplified handler
-	handler := homebrew.NewHandler()
+	handler := install.NewHandler()
 
 	// Create test matches
 	matches := []operations.FileInput{
 		{
-			PackName:     "dev-tools",
-			RelativePath: "Brewfile",
-			SourcePath:   brewfilePath,
+			PackName:     "testpack",
+			RelativePath: "install.sh",
+			SourcePath:   scriptPath,
 		},
 	}
 
@@ -91,10 +88,9 @@ cask "visual-studio-code"
 	// Verify operation
 	op := ops[0]
 	assert.Equal(t, operations.RunCommand, op.Type)
-	assert.Equal(t, "dev-tools", op.Pack)
-	assert.Equal(t, "homebrew", op.Handler)
-	assert.Contains(t, op.Command, "brew bundle")
-	assert.Contains(t, op.Command, brewfilePath)
+	assert.Equal(t, "testpack", op.Pack)
+	assert.Equal(t, "install", op.Handler)
+	assert.Contains(t, op.Command, scriptPath)
 	assert.NotEmpty(t, op.Sentinel)
 
 	// Test with executor in dry-run mode
@@ -111,23 +107,23 @@ cask "visual-studio-code"
 	assert.Contains(t, results[0].Message, "Would execute")
 }
 
-func TestHomebrewHandler_ExecuteWithDataStore(t *testing.T) {
+func TestInstallHandler_ExecuteWithDataStore(t *testing.T) {
 	// This test verifies actual execution with the datastore
 
-	// Create test Brewfile
+	// Create test script
 	tempDir := t.TempDir()
-	brewfileContent := `brew "git"`
-	brewfilePath := filepath.Join(tempDir, "Brewfile")
-	err := os.WriteFile(brewfilePath, []byte(brewfileContent), 0644)
+	scriptContent := "#!/bin/bash\necho 'Installing test pack'\n"
+	scriptPath := filepath.Join(tempDir, "install.sh")
+	err := os.WriteFile(scriptPath, []byte(scriptContent), 0755)
 	require.NoError(t, err)
 
-	handler := homebrew.NewHandler()
+	handler := install.NewHandler()
 
 	matches := []operations.FileInput{
 		{
-			PackName:     "tools",
-			RelativePath: "Brewfile",
-			SourcePath:   brewfilePath,
+			PackName:     "testpack",
+			RelativePath: "install.sh",
+			SourcePath:   scriptPath,
 		},
 	}
 
@@ -138,9 +134,9 @@ func TestHomebrewHandler_ExecuteWithDataStore(t *testing.T) {
 	store := new(MockSimpleDataStore)
 
 	// Expect RunAndRecord to be called with the correct parameters
-	expectedCommand := fmt.Sprintf("brew bundle --file='%s'", brewfilePath)
-	store.On("RunAndRecord", "tools", "homebrew", expectedCommand, mock.MatchedBy(func(s string) bool {
-		// Sentinel should contain pack, filename and checksum
+	expectedCommand := fmt.Sprintf("bash '%s'", scriptPath)
+	store.On("RunAndRecord", "testpack", "install", expectedCommand, mock.MatchedBy(func(s string) bool {
+		// Sentinel should contain filename and checksum
 		return len(s) > 0
 	})).Return(nil)
 
@@ -159,9 +155,48 @@ func TestHomebrewHandler_ExecuteWithDataStore(t *testing.T) {
 	store.AssertExpectations(t)
 }
 
-func TestHomebrewHandler_ClearIntegration(t *testing.T) {
+func TestInstallHandler_CheckSentinel(t *testing.T) {
+	// Test that CheckSentinel operations work correctly
+
+	handler := install.NewHandler()
+
+	// Create a CheckSentinel operation
+	op := operations.Operation{
+		Type:     operations.CheckSentinel,
+		Pack:     "testpack",
+		Handler:  "install",
+		Sentinel: "install.sh-abc123",
+	}
+
+	// Test when sentinel exists
+	store := new(MockSimpleDataStore)
+
+	store.On("HasSentinel", "testpack", "install", "install.sh-abc123").Return(true, nil)
+
+	executor := operations.NewExecutor(store, nil, false)
+	results, err := executor.Execute([]operations.Operation{op}, handler)
+
+	require.NoError(t, err)
+	assert.Len(t, results, 1)
+	assert.True(t, results[0].Success)
+	assert.Equal(t, "Already completed", results[0].Message)
+
+	// Test when sentinel doesn't exist
+	store2 := new(MockSimpleDataStore)
+	store2.On("HasSentinel", "testpack", "install", "install.sh-abc123").Return(false, nil)
+
+	executor2 := operations.NewExecutor(store2, nil, false)
+	results2, err := executor2.Execute([]operations.Operation{op}, handler)
+
+	require.NoError(t, err)
+	assert.Len(t, results2, 1)
+	assert.True(t, results2[0].Success)
+	assert.Equal(t, "Not completed", results2[0].Message)
+}
+
+func TestInstallHandler_ClearIntegration(t *testing.T) {
 	// Test clear functionality
-	handler := homebrew.NewHandler()
+	handler := install.NewHandler()
 
 	// Create mock store and executor
 	store := new(MockSimpleDataStore)
@@ -170,7 +205,7 @@ func TestHomebrewHandler_ClearIntegration(t *testing.T) {
 	// Clear context
 	ctx := operations.ClearContext{
 		Pack: types.Pack{
-			Name: "dev-tools",
+			Name: "testpack",
 		},
 		DryRun: true,
 	}
@@ -182,82 +217,36 @@ func TestHomebrewHandler_ClearIntegration(t *testing.T) {
 
 	// Check cleared item
 	item := clearedItems[0]
-	assert.Equal(t, "homebrew_state", item.Type)
-	assert.Contains(t, item.Path, "dev-tools/homebrew")
-	assert.Contains(t, item.Description, "Would remove Homebrew state")
-	assert.Contains(t, item.Description, "DODOT_HOMEBREW_UNINSTALL=true")
+	assert.Equal(t, "provision_state", item.Type)
+	assert.Contains(t, item.Path, "testpack/install")
+	assert.Equal(t, "Would remove install run records", item.Description)
 }
 
-func TestHomebrewHandler_ClearWithUninstall(t *testing.T) {
-	// Test clear with uninstall enabled
-	_ = os.Setenv("DODOT_HOMEBREW_UNINSTALL", "true")
-	defer func() { _ = os.Unsetenv("DODOT_HOMEBREW_UNINSTALL") }()
+func TestInstallHandler_IdempotentExecution(t *testing.T) {
+	// Test that scripts with same content get same sentinel
 
-	handler := homebrew.NewHandler()
-
-	// Create mock store and executor
-	store := new(MockSimpleDataStore)
-	executor := operations.NewExecutor(store, nil, false)
-
-	// Clear context
-	ctx := operations.ClearContext{
-		Pack: types.Pack{
-			Name: "dev-tools",
-		},
-		DryRun: true,
-	}
-
-	// Execute clear
-	clearedItems, err := executor.ExecuteClear(handler, ctx)
-	require.NoError(t, err)
-	assert.Len(t, clearedItems, 1)
-
-	// Check cleared item reflects uninstall
-	item := clearedItems[0]
-	assert.Contains(t, item.Description, "Would uninstall Homebrew packages")
-}
-
-func TestHomebrewHandler_MultipleBrewfiles(t *testing.T) {
-	// Test handling multiple Brewfiles in different directories
+	// Create test script
 	tempDir := t.TempDir()
-
-	// Create multiple Brewfiles
-	brewfile1 := filepath.Join(tempDir, "Brewfile")
-	err := os.WriteFile(brewfile1, []byte("brew \"git\"\n"), 0644)
+	scriptContent := "#!/bin/bash\necho 'test'\n"
+	scriptPath := filepath.Join(tempDir, "install.sh")
+	err := os.WriteFile(scriptPath, []byte(scriptContent), 0755)
 	require.NoError(t, err)
 
-	appsDir := filepath.Join(tempDir, "apps")
-	err = os.MkdirAll(appsDir, 0755)
-	require.NoError(t, err)
+	handler := install.NewHandler()
 
-	brewfile2 := filepath.Join(appsDir, "Brewfile")
-	err = os.WriteFile(brewfile2, []byte("cask \"slack\"\n"), 0644)
-	require.NoError(t, err)
-
-	handler := homebrew.NewHandler()
-
-	matches := []operations.FileInput{
-		{
-			PackName:     "tools",
-			RelativePath: "Brewfile",
-			SourcePath:   brewfile1,
-		},
-		{
-			PackName:     "tools",
-			RelativePath: "apps/Brewfile",
-			SourcePath:   brewfile2,
-		},
+	match := operations.FileInput{
+		PackName:     "test",
+		RelativePath: "install.sh",
+		SourcePath:   scriptPath,
 	}
 
-	ops, err := handler.ToOperations(matches)
+	// Generate operations twice
+	ops1, err := handler.ToOperations([]operations.FileInput{match})
 	require.NoError(t, err)
-	assert.Len(t, ops, 2)
 
-	// Each should have unique sentinel
-	assert.NotEqual(t, ops[0].Sentinel, ops[1].Sentinel)
+	ops2, err := handler.ToOperations([]operations.FileInput{match})
+	require.NoError(t, err)
 
-	// Both should use brew bundle
-	for _, op := range ops {
-		assert.Contains(t, op.Command, "brew bundle")
-	}
+	// Same content should produce same sentinel
+	assert.Equal(t, ops1[0].Sentinel, ops2[0].Sentinel)
 }
