+++ pkg/handlers/lib/path/path_clear_test.go
@@ -12,7 +12,7 @@ package path_test
 import (
 	"testing"
 
-	"github.com/arthur-debert/dodot/pkg/handlers/path"
+	"github.com/arthur-debert/dodot/pkg/handlers/lib/path"
 	"github.com/arthur-debert/dodot/pkg/testutil"
 	"github.com/arthur-debert/dodot/pkg/types"
 	"github.com/stretchr/testify/assert"
