+++ pkg/handlers/lib/shell/integration_test.go
@@ -1,10 +1,9 @@
-package symlink_test
+package shell_test
 
 import (
-	"os"
 	"testing"
 
-	"github.com/arthur-debert/dodot/pkg/handlers/symlink"
+	"github.com/arthur-debert/dodot/pkg/handlers/lib/shell"
 	"github.com/arthur-debert/dodot/pkg/operations"
 	"github.com/arthur-debert/dodot/pkg/types"
 	"github.com/stretchr/testify/assert"
@@ -56,41 +55,43 @@ func (m *MockSimpleDataStore) ListHandlerSentinels(pack, handlerName string) ([]
 	return args.Get(0).([]string), args.Error(1)
 }
 
-func TestSymlinkHandler_OperationIntegration(t *testing.T) {
-	// This test verifies the symlink handler works with the operation system
-
-	// Set HOME for consistent tests
-	oldHome := os.Getenv("HOME")
-	_ = os.Setenv("HOME", "/home/testuser")
-	defer func() { _ = os.Setenv("HOME", oldHome) }()
+func TestShellHandler_OperationIntegration(t *testing.T) {
+	// This test verifies the shell handler works with the operation system
 
 	// Create simplified handler
-	handler := symlink.NewHandler()
+	handler := shell.NewHandler()
 
 	// Create test matches
 	matches := []operations.FileInput{
 		{
-			PackName:     "vim",
-			RelativePath: ".vimrc",
-			SourcePath:   "/dotfiles/vim/.vimrc",
+			PackName:     "bash",
+			RelativePath: "aliases.sh",
+			SourcePath:   "/dotfiles/bash/aliases.sh",
 		},
 		{
-			PackName:     "vim",
-			RelativePath: ".vim/colors/theme.vim",
-			SourcePath:   "/dotfiles/vim/.vim/colors/theme.vim",
+			PackName:     "bash",
+			RelativePath: "functions.sh",
+			SourcePath:   "/dotfiles/bash/functions.sh",
+		},
+		{
+			PackName:     "zsh",
+			RelativePath: "config.zsh",
+			SourcePath:   "/dotfiles/zsh/config.zsh",
 		},
 	}
 
 	// Convert to operations
 	ops, err := handler.ToOperations(matches)
 	require.NoError(t, err)
-	assert.Len(t, ops, 4) // 2 operations per file
+	assert.Len(t, ops, 3) // One operation per script
 
 	// Verify operations
-	assert.Equal(t, operations.CreateDataLink, ops[0].Type)
-	assert.Equal(t, operations.CreateUserLink, ops[1].Type)
-	assert.Equal(t, operations.CreateDataLink, ops[2].Type)
-	assert.Equal(t, operations.CreateUserLink, ops[3].Type)
+	for i, op := range ops {
+		assert.Equal(t, operations.CreateDataLink, op.Type)
+		assert.Equal(t, "shell", op.Handler)
+		assert.Equal(t, matches[i].PackName, op.Pack)
+		assert.Equal(t, matches[i].SourcePath, op.Source)
+	}
 
 	// Test with executor in dry-run mode
 	store := new(MockSimpleDataStore)
@@ -99,30 +100,25 @@ func TestSymlinkHandler_OperationIntegration(t *testing.T) {
 	// Execute operations
 	results, err := executor.Execute(ops, handler)
 	require.NoError(t, err)
-	assert.Len(t, results, 4)
+	assert.Len(t, results, 3)
 
 	// All should be successful in dry run
 	for _, result := range results {
 		assert.True(t, result.Success)
-		assert.Contains(t, result.Message, "Would")
+		assert.Contains(t, result.Message, "Would create data link")
 	}
 }
 
-func TestSymlinkHandler_ExecuteWithDataStore(t *testing.T) {
+func TestShellHandler_ExecuteWithDataStore(t *testing.T) {
 	// This test verifies actual execution with the datastore
 
-	// Set HOME for consistent tests
-	oldHome := os.Getenv("HOME")
-	_ = os.Setenv("HOME", "/home/testuser")
-	defer func() { _ = os.Setenv("HOME", oldHome) }()
-
-	handler := symlink.NewHandler()
+	handler := shell.NewHandler()
 
 	matches := []operations.FileInput{
 		{
-			PackName:     "vim",
-			RelativePath: ".vimrc",
-			SourcePath:   "/dotfiles/vim/.vimrc",
+			PackName:     "bash",
+			RelativePath: "aliases.sh",
+			SourcePath:   "/dotfiles/bash/aliases.sh",
 		},
 	}
 
@@ -133,36 +129,27 @@ func TestSymlinkHandler_ExecuteWithDataStore(t *testing.T) {
 	store := new(MockSimpleDataStore)
 
 	// Expect CreateDataLink to be called
-	store.On("CreateDataLink", "vim", "symlink", "/dotfiles/vim/.vimrc").
-		Return("/datastore/vim/symlinks/.vimrc", nil)
-
-	// Expect CreateUserLink to be called
-	// NOTE: In the current implementation, the executor passes op.Source
-	// directly rather than the datastore path. This is a known issue
-	// that will be resolved in Phase 3 when we remove the adapters.
-	store.On("CreateUserLink", "/dotfiles/vim/.vimrc", "/home/testuser/.vimrc").
-		Return(nil)
+	store.On("CreateDataLink", "bash", "shell", "/dotfiles/bash/aliases.sh").
+		Return("/datastore/bash/shell/aliases.sh", nil)
 
 	// Execute with real mode (not dry-run)
 	executor := operations.NewExecutor(store, nil, false)
 	results, err := executor.Execute(ops, handler)
 
 	require.NoError(t, err)
-	assert.Len(t, results, 2)
+	assert.Len(t, results, 1)
 
-	// Both operations should succeed
+	// Operation should succeed
 	assert.True(t, results[0].Success)
 	assert.Contains(t, results[0].Message, "Created data link")
-	assert.True(t, results[1].Success)
-	assert.Contains(t, results[1].Message, "Created link")
 
 	// Verify all expectations were met
 	store.AssertExpectations(t)
 }
 
-func TestSymlinkHandler_Clear(t *testing.T) {
+func TestShellHandler_ClearIntegration(t *testing.T) {
 	// Test clear functionality
-	handler := symlink.NewHandler()
+	handler := shell.NewHandler()
 
 	// Create mock store and executor
 	store := new(MockSimpleDataStore)
@@ -171,7 +158,7 @@ func TestSymlinkHandler_Clear(t *testing.T) {
 	// Clear context
 	ctx := operations.ClearContext{
 		Pack: types.Pack{
-			Name: "vim",
+			Name: "bash",
 		},
 		DryRun: true,
 	}
@@ -183,7 +170,7 @@ func TestSymlinkHandler_Clear(t *testing.T) {
 
 	// Check cleared item
 	item := clearedItems[0]
-	assert.Equal(t, "symlink_state", item.Type)
-	assert.Contains(t, item.Path, "vim/symlink")
-	assert.Contains(t, item.Description, "Would remove symlink")
+	assert.Equal(t, "shell_state", item.Type)
+	assert.Contains(t, item.Path, "bash/shell")
+	assert.Equal(t, "Would remove shell state", item.Description)
 }
