+++ pkg/handlers/pipeline/integration.go
@@ -11,11 +11,11 @@ import (
 	"github.com/arthur-debert/dodot/pkg/errors"
 	exec "github.com/arthur-debert/dodot/pkg/execution"
 	"github.com/arthur-debert/dodot/pkg/handlers"
-	"github.com/arthur-debert/dodot/pkg/handlers/homebrew"
-	"github.com/arthur-debert/dodot/pkg/handlers/install"
-	"github.com/arthur-debert/dodot/pkg/handlers/path"
-	"github.com/arthur-debert/dodot/pkg/handlers/shell"
-	"github.com/arthur-debert/dodot/pkg/handlers/symlink"
+	"github.com/arthur-debert/dodot/pkg/handlers/lib/homebrew"
+	"github.com/arthur-debert/dodot/pkg/handlers/lib/install"
+	"github.com/arthur-debert/dodot/pkg/handlers/lib/path"
+	"github.com/arthur-debert/dodot/pkg/handlers/lib/shell"
+	"github.com/arthur-debert/dodot/pkg/handlers/lib/symlink"
 	"github.com/arthur-debert/dodot/pkg/logging"
 	"github.com/arthur-debert/dodot/pkg/operations"
 	"github.com/arthur-debert/dodot/pkg/types"
